flask/requirements.txt
-i https://pypi.org/simple
absl-py==1.3.0 ; python_version >= '3.6'
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; sys_platform == 'win32'
contourpy==1.0.6 ; python_version >= '3.7'
cycler==0.11.0 ; python_version >= '3.6'
decorator==5.1.1 ; python_version >= '3.5'
executing==1.2.0
flask==2.2.2
flask-cors==3.0.10
flask-mysqldb==1.0.1
flatbuffers==1.12
fonttools==4.38.0 ; python_version >= '3.7'
gast==0.4.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth==2.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==0.4.6 ; python_version >= '3.6'
google-pasta==0.2.0
grpcio==1.51.1 ; python_version >= '3.7'
h5py==3.7.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
ipython==8.7.0
itsdangerous==2.1.2 ; python_version >= '3.7'
jedi==0.18.2 ; python_version >= '3.6'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
keras==2.9.0
keras-preprocessing==1.1.2
kiwisolver==1.4.4 ; python_version >= '3.7'
libclang==14.0.6
markdown==3.4.1 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
matplotlib==3.6.2 ; python_version >= '3.8'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
mysqlclient==2.1.1 ; python_version >= '3.5'
nltk==3.8
numpy==1.24.1
oauthlib==3.2.2 ; python_version >= '3.6'
opencv-python==4.7.0.68
opt-einsum==3.3.0 ; python_version >= '3.5'
packaging==22.0 ; python_version >= '3.7'
pandas==1.5.2
parso==0.8.3 ; python_version >= '3.6'
pickleshare==0.7.5
pillow==9.3.0 ; python_version >= '3.7'
prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'
protobuf==3.19.6 ; python_version >= '3.5'
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.2.3
pygments==2.14.0 ; python_version >= '3.6'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7
regex==2022.10.31 ; python_version >= '3.6'
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
seaborn==0.12.2
setuptools==65.6.3 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.6.2
tensorboard==2.9.1 ; python_version >= '3.6'
tensorboard-data-server==0.6.1 ; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0 ; python_version >= '3.7'
tensorflow-io-gcs-filesystem==0.29.0 ; python_version < '3.12' and python_version >= '3.7'
termcolor==2.1.1 ; python_version >= '3.7'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.8.0 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wcwidth==0.2.5
werkzeug==2.2.2 ; python_version >= '3.7'
wheel==0.38.4 ; python_version >= '3.7'
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
